Allow initializing pages with content
If the page has a list of items of interest, it
might be easier to write a scraper that creates
a new page with a HTML fragment from an existing
page.

We still have to parse a URL in if we want to
follow links off (in case the links in the page
are relative).